Refactor to prefer composition
* Restructured modules to avoid long files
* Created abc's for main extension classes
* Added integration tests for sync and twisted
* Removed inheritence-based twisted extension
